split multi-class files into more files
