+++ src/Keda.Samples.DotNet.Web/Program.cs
@@ -24,7 +24,7 @@ builder.Services.AddOrderQueueServices();
 
 
 var app = builder.Build();
-if (app.Environment.IsDevelopment())app.UseDeveloperExceptionPage();
+if (app.Environment.IsDevelopment()) app.UseDeveloperExceptionPage();
 else app.UseExceptionHandler("/Error");
 
 app.UseStaticFiles();
