+++ src/Keda.Samples.Dotnet.Contracts/Customer.cs
@@ -1,0 +1,3 @@
+namespace Keda.Samples.Dotnet.Contracts;
+
+public record Customer(string FirstName, string LastName);
