+++ src/Keda.Samples.Dotnet.Contracts/Order.cs
@@ -1,0 +1,4 @@
+namespace Keda.Samples.Dotnet.Contracts;
+
+
+public record Order(string Id, int Amount, string ArticleNumber, Customer Customer);
