+++ src/Keda.Samples.Dotnet.Contracts/QueueStatus.cs
@@ -1,0 +1,3 @@
+﻿namespace Keda.Samples.Dotnet.Contracts;
+
+public record QueueStatus(long MessageCount);
